requirements.txt
python-dotenv
pykakasi
dearpygui
customtkinter
